third_party/blink/web_tests/external/wpt/paint-timing/fcp-only/fcp-svg-expected.txt

This is a testharness.js-based test.
FAIL First contentful paint fires when SVG becomes contentful. assert_equals: First contentful paint marked too early.  expected 0 but got 1
Harness: the test ran to completion.
